+++ examples/ocr2/src/main.rs
@@ -17,7 +17,7 @@ extern crate simplelog;
 extern crate darwin_rs;
 
 use std::sync::Arc;
-use rand::{Rng, sample};
+use rand::Rng;
 use std::fs::File;
 use std::io::Read;
 use std::path::Path;
@@ -27,7 +27,7 @@ use simplelog::{SimpleLogger, LogLevelFilter};
 use std::str;
 
 // internal modules
-use darwin_rs::{Individual, SimulationBuilder, Population, PopulationBuilder, SimError};
+use darwin_rs::{Individual, SimulationBuilder, Population, PopulationBuilder, simulation_builder};
 
 const MIN_ASCII: u8 = 32;
 const MAX_ASCII: u8 = 126;
@@ -55,20 +55,17 @@ fn make_all_populations1<'a>(individuals: u32, config: &OCRConfig<'a>, populatio
 
     let initial_population = make_population(individuals, &config);
 
-    let mut rng = rand::thread_rng();
-
     for i in 1..(populations + 1) {
 
         let mut pop = PopulationBuilder::<OCRItem>::new()
             .set_id(i)
             .initial_population(&initial_population)
-            .mutation_rate(sample(&mut rng, 1..100000, individuals as usize));
-//            .mutation_rate((1000..7000).cycle().take(individuals as usize).collect());
+            .increasing_exp_mutation_rate(1.01);
 
         if i == populations {
             // Special case for the last popilation
             pop = pop.reset_limit_start(100);
-            pop = pop.reset_limit_end(1000);
+            pop = pop.reset_limit_end(10000);
             pop = pop.reset_limit_increment(100);
         } else {
             pop = pop.reset_limit_end(0);
@@ -310,16 +307,21 @@ fn draw_text_line(canvas: &mut ImageBuffer<Luma<u8>, Vec<u8>>,
                 let y = (y as i32) + bb.min.y + pos_y;
                 if x >=0 && y >= 0 && x < canvas.width() as i32 && y < canvas.height() as i32 {
                     canvas.put_pixel(x as u32, y as u32, Luma::<u8>{ data: [(v * 255.0) as u8] } );
-                } else { result = false; }
+                } else {
+                    result = false;
+                    return;
+                }
             })
         }
+
+        if !result { break }
     }
 
     result
 }
 
 fn main() {
-    println!("Darwin test: optical character recognition");
+    println!("Darwin test: optical character recognition 2");
 
     let _ = SimpleLogger::init(LogLevelFilter::Info);
 
@@ -348,8 +350,8 @@ fn main() {
     draw_text_line(&mut original_img, &font_config, 10, 10, "Darwin-rs");
     draw_text_line(&mut original_img, &font_config, 10, 40, "OCR Test!");
 
-//    let img_file = Path::new("rendered_text.png");
-//    let _ = original_img.save(&img_file);
+    let img_file = Path::new("rendered_text.png");
+    let _ = original_img.save(&img_file);
 
     let ocr_config = OCRConfig {
             font_config: font_config,
@@ -357,16 +359,18 @@ fn main() {
         };
 
     let num_populations = 16;
+    let num_of_individuals = 100;
 
     let ocr_builder = SimulationBuilder::<OCRItem>::new()
         .fitness(0.0)
         .threads(16)
-        .add_multiple_populations(make_all_populations1(16, &ocr_config, num_populations as u32))
+        .add_multiple_populations(make_all_populations1(num_of_individuals, &ocr_config, num_populations as u32))
         .share_fittest()
         .finalize();
 
     match ocr_builder {
-        Err(SimError::EndIterationTooLow) => println!("more than 10 iteratons needed"),
+        Err(simulation_builder::Error(simulation_builder::ErrorKind::EndIterationTooLow, _)) => println!("more than 10 iteratons needed"),
+        Err(e) => println!("unexpected error: {}", e),
         Ok(mut ocr_simulation) => {
             ocr_simulation.run();