+++ src/simulation_builder.rs
@@ -2,7 +2,7 @@
 //!
 //! darwin-rs: evolutionary algorithms with Rust
 //!
-//! Written by Willi Kappler, Version 0.3 (2016.08.29)
+//! Written by Willi Kappler, Version 0.4 (2017.??)
 //!
 //! Repository: https://github.com/willi-kappler/darwin-rs
 //!
@@ -28,15 +28,11 @@ pub struct SimulationBuilder<T: Individual + Send + Sync> {
     simulation: Simulation<T>,
 }
 
-quick_error! {
-    #[derive(Debug)]
-    pub enum SimError {
-        /// The number of iteration is too low, should be >= 10
-        EndIterationTooLow {}
+error_chain! {
+    errors {
+        EndIterationTooLow
     }
 }
-
-pub type Result<T> = std::result::Result<Simulation<T>, SimError>;
 
 /// This implementation contains all the helper method to build (configure) a valid simulation.
 impl<T: Individual + Send + Sync> SimulationBuilder<T> {
@@ -55,7 +51,12 @@ impl<T: Individual + Send + Sync> SimulationBuilder<T> {
                     fittest: Vec::new(),
                     iteration_counter: 0
                 },
-                share_fittest: false
+                share_fittest: false,
+                num_of_global_fittest: 10,
+                output_every: 10,
+                output_every_counter: 0,
+                share_every: 10,
+                share_counter: 0
             },
         }
     }
@@ -108,12 +109,32 @@ impl<T: Individual + Send + Sync> SimulationBuilder<T> {
         self
     }
 
+    /// How many global fittest should be kept ? (The size of the "high score list")
+    pub fn num_of_global_fittest(mut self, num_of_global_fittest: usize) -> SimulationBuilder<T> {
+        self.simulation.num_of_global_fittest = num_of_global_fittest;
+        self
+    }
+
+    /// Do not output every time a new individual is found, only every nth time.
+    /// n == output_every
+    pub fn output_every(mut self, output_every: u32) -> SimulationBuilder<T> {
+        self.simulation.output_every = output_every;
+        self
+    }
+
+    /// If share fittest is enabled and the number share_every of iteration has passed then
+    /// the fittest individual is shared between all populations
+    pub fn share_every(mut self, share_every: u32) -> SimulationBuilder<T> {
+        self.simulation.share_every = share_every;
+        self
+    }
+
     /// This checks the configuration of the simulation and returns an error or Ok if no errors
     /// where found.
-    pub fn finalize(self) -> Result<T> {
+    pub fn finalize(self) -> Result<Simulation<T>> {
         match self.simulation {
             Simulation { type_of_simulation: SimulationType::EndIteration(0...9), .. } => {
-                Err(SimError::EndIterationTooLow)
+                Err(ErrorKind::EndIterationTooLow.into())
             }
             _ => Ok(self.simulation),
         }